#14 Rewrite/restore exposed services on maintenance mode switching
Exposed services that provide direct pod access cannot be re-routed with
ingress objects. Yet, all requests to dogus should be sensibly rejected during
the maintenance mode.

This commit rewrites the service selector of each exposed service, so it
receives a wrong target (and thus returns an HTTP error). Upon deactivation the
exposed service will be restored by overwriting the selector with the original
selector, that is: with the dogu name label.