[#88] Implement redirector.go to create redirect alternative fqdns
The Redirector creates a single ingress resource which redirects hosts mentioned in alternative fqnds to the primary fqdn. Although redirection is done the ingress controller needs some services defined in the rule paths. For this reason ces-loadbalancer has been used as a static resource.

Moreover, nginx ingress annotation nginx.ingress.kubernetes.io/server-snippet has been used for the redirect as we need to preserve the full path for the redirect. In the current version of nginx ingress controller (v1.12.1) it is not allowed to use the annotation nginx.ingress.kubernetes.io/permanent-redirect with "$request_uri".